+++ packages/core/index.ts
@@ -5,7 +5,11 @@ export { ModalCtrl } from './src/controllers/ModalCtrl'
 export { OptionsCtrl } from './src/controllers/OptionsCtrl'
 export { RouterCtrl } from './src/controllers/RouterCtrl'
 export { ToastCtrl } from './src/controllers/ToastCtrl'
-export type { Listing, ListingResponse, RouterView } from './src/types/controllerTypes'
-export type { Settings } from './src/types/settingsTypes'
+export type {
+  ConfigCtrlState,
+  Listing,
+  ListingResponse,
+  RouterView
+} from './src/types/controllerTypes'
 export { CoreHelpers } from './src/utils/CoreHelpers'
 import './src/utils/PolyfillUtil'
