+++ packages/core/src/controllers/ConfigCtrl.ts
@@ -1,6 +1,5 @@
 import { proxy, subscribe as valtioSub } from 'valtio/vanilla'
 import type { ConfigCtrlState } from '../types/controllerTypes'
-import type { Settings } from '../types/settingsTypes'
 import { OptionsCtrl } from './OptionsCtrl'
 
 // -- initial state ------------------------------------------------ //
@@ -9,7 +8,6 @@ function isDarkMode() {
 }
 
 const state = proxy<ConfigCtrlState>({
-  configured: false,
   projectId: undefined,
   theme: isDarkMode() ? 'dark' : 'light',
   accentColor: 'default',
@@ -24,9 +22,8 @@ export const ConfigCtrl = {
     return valtioSub(state, () => callback(state))
   },
 
-  setConfig(config: Settings) {
+  setConfig(config: ConfigCtrlState) {
     if (config.standaloneChains?.length) OptionsCtrl.setStandaloneChains(config.standaloneChains)
     Object.assign(state, config)
-    state.configured = true
   }
 }
