+++ packages/core/src/types/controllerTypes.ts
@@ -1,10 +1,39 @@
 import type { EthereumClient } from '@web3modal/ethereum'
-import type { Settings } from './settingsTypes'
+
+interface MobileWallet {
+  id: string
+  links: {
+    universal: string
+    deep?: string
+  }
+}
+
+interface DesktopWallet {
+  id: string
+  links: {
+    deep: string
+    universal: string
+  }
+}
 
 // -- ConfigCtrl ------------------------------------------- //
-
-export interface ConfigCtrlState extends Settings {
-  configured: boolean
+export interface ConfigCtrlState {
+  projectId?: string
+  theme?: 'dark' | 'light'
+  accentColor?:
+    | 'blackWhite'
+    | 'blue'
+    | 'default'
+    | 'green'
+    | 'magenta'
+    | 'orange'
+    | 'purple'
+    | 'teal'
+  standaloneChains?: string[]
+  mobileWallets?: MobileWallet[]
+  desktopWallets?: DesktopWallet[]
+  walletImages?: Record<string, string>
+  chainImages?: Record<string, string>
 }
 
 // -- ModalCtrl --------------------------------------- //
